TEXT!

+++ src/css/cssnodes/values/color.cpp
@@ -5,7 +5,7 @@ void Color::display(){
   printf("Color {%d %d %d %d}", r,g,b,a);
 }
 
-float Color::toPX() {
+int Color::toPX() {
   std::cout << "Cannot convert color to PX!\n";
   return 0;
 }
