+++ src/css/cssnodes/values/value.hpp
@@ -5,6 +5,6 @@ class Value {
   public:
     virtual ~Value()=0;
     virtual void display();
-    virtual float toPX();
+    virtual int toPX();
 };
 #endif
